Add: hide/enable offer button in company offers management page, in mobile view
